Setup the HTML structure I wanted for this and have started styling in CSS
I have four main divs plus a footer for the HTML structure all with sub divs as needed. Ive set the body to flex with flex-direction column, and then set the header class also to display: flex as well to then seperate the logo and the nav bar links eventually.