I got my main div proportions setup atleast lol
My edits on this are all over the place and I honestly wanted to just completely start over once I realized some of the stuff I have is stupid for individualized elements or class selectors in CSS but the biggest changes that I'm actually happy with was actually getting the body to show 100% of the screen when looking at edits in the dev tools in the browser and then resetting margins and padding to just do all of that myself with the univsersal selectors. Also I added a "box" class to all the main divs for the four main portions of the page so I can univserally add padding and dimensions to each instead of each individualized div. I am happy with that atleast even though it felt like I did not accomplish much past that.