put the styles I like in for the .container-2 and .container-3
That includes some of the text for the fictious raccoon mercenary company. All of the sub divs have been put into place as well, this inlcudes adding a class tag to the img elements in .container-3 which I think wasnt actually necassary if I just used desecendant combinators to select them with the above div, no harm in it though I suppose.