Added the picture files to an img folder in the main directory
Also have the footer and the call to action divs compeleted and stylized as per the design outcome picture. also replaced the dummy links with picture links to finish the design as well as the rest of the dummy text for the fake company. I should go through and edit the html and css to add comments to say what each section is for......